Large Scale Simulated Spatial Multisampling Data (DataFrame)

A large scale simulation data of spatial hierarchical structure and samples overlapping on certain locations.

Usage

data(multisampling)

Format

A list of three items called "data", "coords" and "beta". Item "data" is a data frame with 13862 observations at 200 locations and the following 6 variables.

y

a numeric vector, dependent variable \(y\)

g1

a numeric vector, group level independent variable \(g_1\)

g2

a numeric vector, group level independent variable \(g_2\)

z1

a numeric vector, sample level independent variable \(z_1\)

x1

a numeric vector, sample level independent variable \(x_1\)

group

a numeric vector, group id of each sample

where g1 and g2 are used to estimate local fixed effects; x1 is used to estimate global fixed effects and z1 is used to estimate random effects.

Author

Yigong Hu yigong.hu@bristol.ac.uk

Examples

if (FALSE) { # \dontrun{
data(multisampling.large)
hgwr(formula = y ~ L(g1 + g2) + x1 + (z1 | group),
     data = multisampling.large$data,
     coords = multisampling.large$coords,
     bw = 32, kernel = "bisquared")
} # }