Print a character matrix as a table.

Print a character matrix as a table.

Usage

# S3 method for class 'table.md'
print(
  x,
  col.sep = "",
  header.sep = "",
  row.begin = "",
  row.end = "",
  table.style = c("plain", "md", "latex"),
  ...
)

Arguments

x

A character matrix.

col.sep

Column separator. Default to "".

header.sep

Header separator. Default to "-".

row.begin

Character at the beginning of each row. Default to col.sep.

row.end

Character at the ending of each row. Default to col.sep.

table.style

Name of pre-defined style. Possible values are "plain", "md" or "latex". Default to "plain".

...

Additional style control arguments.

Value

No return.

Details

When table.style is specified, col.sep, header.sep, row.begin and row.end would not take effects. Because this function will automatically set their values. For each possible value of table.style, its corresponding style settings are shown in the following table.

plain	md	latex

col.sep	""	"|"	"&"

header.sep	""	"-"	""

row.begin	""	"|"	""

row.end	""	"|"	"\\"

In this function, characters are right padded by spaces.

See also