Summary an hgwrm object.

Summary an hgwrm object.

Usage

# S3 method for class 'hgwrm'
summary(object, ..., test_hetero = FALSE)

Arguments

object

An hgwrm object returned from hgwr().

...

Other arguments passed from other functions.

test_hetero

Logical/list value. Whether to test the spatial heterogeneity of local fixed effects. If it is set to FALSE, the test will not be executed. If it is set to TRUE, the test will be executed with default parameters (see details below). It accepts a list to enable the test with specified parameters.

Value

A list containing summary informations of this hgwrm object with the following fields.

diagnostic

A list of diagnostic information.

random.stddev

The standard deviation of random effects.

random.corr

The correlation matrix of random effects.

residuals

The residual vector.

Details

The parameters used to perform test of spatial heterogeneity are

bw

Bandwidth (unit: number of nearest neighbours) used to make spatial kernel density estimation. Default: 10.

poly

The number of polynomial terms used in the local polynomial estimation. Default: 2.

resample

Total resampling times. Default: 5000.

See also