Wuhan Second-hand House Price and POI Data (DataFrame)

A data set of second-hand house price in Wuhan, China collected in 2018.

Usage

data(multisampling)

Format

A list of two items called "data" and "coords". Item "data" is a data frame with 13862 second-hand properties at 779 neighbourhoods and the following 22 variables.

Price

House price per square metre.

Floor.High

1 if a property is on a high floor, otherwise 0.

Floor.Low

1 if a property is on a low floor, otherwise 0.

Decoration.Fine

1 if a property is well decorated, otherwise 0.

PlateTower

1 if a property is of the plate-tower type, otherwise 0.

Steel

1 if a property is of 'steel' structure, otherwise 0.

BuildingArea

Building area in square metres.

Fee

Management fee per square meter per month.

d.Commercial

Distance to the nearest commercial area.

d.Greenland

Distance to the nearest green land.

d.Water

Distance to the nearest river or lake.

d.University

Distance to the nearest university.

d.HighSchool

Distance to the nearest high school.

d.MiddleSchool

Distance to the nearest middle school.

d.PrimarySchool

Distance to the nearest primary school.

d.Kindergarten

Distance to the nearest kindergarten.

d.SubwayStation

Distance to the nearest subway station.

d.Supermarket

Distance to the nearest supermarket.

d.ShoppingMall

Distance to the nearest shopping mall.

lon

Longitude coordinates (Projected CRS: EPSG 3857).

lat

Latitude coordinates (Projected CRS: EPSE 3857).

group

Group id of each sample.

The following variables are group level:

- Fee - d.Commercial - d.Greenland - d.Water - d.University - d.HighSchool - d.MiddleSchool - d.PrimarySchool - d.Kindergarten - d.SubwayStation - d.Supermarket - d.ShoppingMall

The follwoing variables are sample level:

- Price - Floor.High - Floor.Low - Decoration.Fine - PlateTower - Steel - BuildingArea

Item "coords" is a 779-by-2 matrix of coordinates of all neighbourhoods.

Author

Yigong Hu yigong.hu@bristol.ac.uk

Examples

if (FALSE) {
data(wuhan.hp)
hgwr(
  formula = Price ~ L(d.Water + d.Commercial + d.PrimarySchool +
            d.Kindergarten + Fee) + BuildingArea + (Floor.High | group),
  data = wuhan.hp$data,
  coords = wuhan.hp$coords, bw = 50, kernel = "bisquared")
}